+++ application/models/FleetModel.php
@@ -4,6 +4,8 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class FleetModel extends CSV_Model
 {
+    private $budget = 10000;
+    private $plane_types = null;
 
     // Constructor
     public function __construct()
@@ -82,6 +84,68 @@ class FleetModel extends CSV_Model
             ['field' => 'model_id', 'label' => 'Model Id', 'rules' => 'required'],
         );
             return $config;
+    }
+    
+    /**
+     * Retrieve the total budget allocated to this fleet.
+     * @return type double
+     */
+    public function getBudget(){
+        return doubleval($this->budget);
+    }
+    
+    /**
+     * Return how much money has been spent in buying planes for this fleet.
+     * @return type double
+     */
+    public function getGrandTotal(){
+        $wackyModel = new WackyModel();
+        $planes_raw = $wackyModel->airplanes();
+        $planes = array();
+        foreach($planes_raw as $key=>$plane){
+            $planes[$plane['id']] = $plane;
+        }
+        $grandTotal = 0.0;
+        foreach ($this->all() as $key => $value) {
+            $grandTotal += doubleval($planes[$value->model_id]['price']);
+        }
+        return $grandTotal;
+    }
+    
+    /**
+     * Check if the given plane type is valid
+     * @param type $type
+     */
+    public function checkPlaneType($type){
+        if($this->plane_types == null){
+            $wackyModel = new WackyModel();
+            $planes = $wackyModel->airplanes();
+            $this->plane_types = array();
+            foreach ($planes as $key => $plane) {
+                $this->plane_types[$plane['id']] = $plane['id'];
+            }
+            return in_array($type, $this->plane_types);
+        }
+    }
+    
+    /**
+     * Check if the budget is sufficient to make a purchase of certain price
+     * @param type $price price
+     */
+    public function checkSufficientBudget($price){
+        return $this->getBudget() - $this->getGrandTotal() - $price >= 0;
+    }
+    
+    /**
+     * A set of rules for form validation.
+     */
+    public function formRules(){
+        //Form Validation from Codeigniter helper
+        $config = array(
+            ['field' => 'plane_id', 'label' => 'Plane Id', 'rules' => 'alpha_numeric|greater_than[0]'],
+            ['field' => 'model_id', 'label' => 'Model', 'rules' => 'required|callback_validateModelId'],
+        );
+        return $config;
     }
     // retrieve a single plane, null if not found
 //    public function get($which);