+++ application/views/booking_header.php
@@ -1,0 +1,12 @@
+<h3>{stops}</h3>
+<table class="table">
+    <tr>
+        <th>Flight ID</th>
+        <th>Plane</th>
+        <th>Departure Airport</th>
+        <th>Departure Time</th>
+        <th>Arrival Airport</th>
+        <th>Arrival Time</th>
+    </tr>
+    {flightoptions}
+</table>