+++ application/views/template.php
@@ -4,7 +4,8 @@
     <title>{pagetitle}</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <!-- Bootstrap -->
-    <link href="/assets/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <!--<link href="/assets/css/bootstrap.min.css" rel="stylesheet">-->
     <link href="/assets/css/style.css" rel="stylesheet">
   </head>
   <body>